+++ test/ethers/tx_data_test.exs
@@ -17,33 +17,33 @@ defmodule Ethers.TxDataTest do
 
   describe "to_map/2" do
     test "converts a TxData to transaction map" do
-      tx_data = TxData.new("0xffff", @function_selector, nil)
+      tx_data = TxData.new("0xffff", @function_selector, nil, nil)
 
       assert %{data: "0xffff"} == TxData.to_map(tx_data, [])
     end
 
     test "includes the default address if any" do
-      tx_data = TxData.new("0xffff", @function_selector, "0xdefault")
+      tx_data = TxData.new("0xffff", @function_selector, "0xdefault", nil)
 
       assert %{data: "0xffff", to: "0xdefault"} == TxData.to_map(tx_data, [])
     end
 
     test "includes overrides in transaction map" do
-      tx_data = TxData.new("0xffff", @function_selector, "0xdefault")
+      tx_data = TxData.new("0xffff", @function_selector, "0xdefault", nil)
 
       assert %{data: "0xffff", to: "0xdefault", from: "0xfrom"} ==
                TxData.to_map(tx_data, from: "0xfrom")
     end
 
     test "can override default address" do
-      tx_data = TxData.new("0xffff", @function_selector, "0xdefault")
+      tx_data = TxData.new("0xffff", @function_selector, "0xdefault", nil)
 
       assert %{data: "0xffff", to: "0xnotdefault"} ==
                TxData.to_map(tx_data, to: "0xnotdefault")
     end
 
     test "integer overrides are converted to hex" do
-      tx_data = TxData.new("0xffff", @function_selector, nil)
+      tx_data = TxData.new("0xffff", @function_selector, nil, nil)
 
       assert %{data: "0xffff", gas: "0x1"} ==
                TxData.to_map(tx_data, gas: 1)