+++ test/support/contracts/revert.sol
@@ -6,4 +6,20 @@ contract Revert {
         require(success, "success must be true");
         return true;
     }
+
+    function reverting() public pure {
+        revert("revert message");
+    }
+
+    error RevertError();
+    error RevertWithMessage(string message);
+    error RevertWithUnnamedArg(uint8);
+
+    function revertingWithError() public pure {
+        revert RevertError();
+    }
+
+    function revertingWithMessage() public pure {
+        revert RevertWithMessage("this is sad!");
+    }
 }